instances conversion to one apge
